starting improvement stuff

+++ server/src/main/java/edu/unh/cs/cs619/bulletzone/model/Playable.java
@@ -114,7 +114,7 @@ public abstract class Playable extends FieldEntity {
         return lastMoveTime;
     }
 
-    public long getAllowedMoveInterval() {
+    public int getAllowedMoveInterval() {
         return allowedMoveInterval;
     }
